Git Automations/requirements.txt
et-xmlfile==1.1.0
gitdb==4.0.10
GitPython==3.1.31
numpy==1.21.6
openpyxl==3.1.2
pandas==1.3.5
pip==22.3.1
python-dateutil==2.8.2
pytz==2023.3
setuptools==65.5.1
six==1.16.0
smmap==5.0.0
typing-extensions==4.6.3
wheel==0.38.4
XlsxWriter==3.1.2
